refactor(AlertEvents): move AlertEvents to src/ui/components [YTFRONT-4144]
