+++ packages/ui/src/ui/components/AlertEvents/AlertEvents.scss
@@ -1,4 +1,4 @@
-.alert-events {
+.yt-alert-events {
     overflow-x: auto;
 
     &__row_current {
